create CalendarRestController with accesible for all request
