+++ src/main/java/com/example/calendar/Controller/ParticipantRestController.java
@@ -32,28 +32,10 @@ import org.springframework.web.bind.annotation.*;
 @CrossOrigin(origins = "*")
 @AllArgsConstructor
 public class ParticipantRestController {
-    
+
     private ParticipantRepository participantRepository;
     private ModelMapper modelMapper;
     private PasswordEncoder encoder;
-
-    @GetMapping
-    public List<ParticipantDto> findAll() {
-        return participantRepository.findParticipantByStatus(ParticipantStatus.ACTIVE).stream().
-                map(participant -> modelMapper.map(participant, ParticipantDto.class)).toList();
-    }
-
-
-    @PostMapping
-    public ParticipantDto addParticipant(@Valid @RequestBody ParticipantFullDto dto) {
-        var participant = dto.toParticipant();
-        participant.setPassword(encoder.encode(participant.getPassword()));
-        if (participantRepository.findParticipantsWithEqualEmailAndNonEqualId(participant.getEmail(), 0L).size() != 0) {
-            throw new EmailNotUnique(participant.getEmail());
-        }
-        return modelMapper.map(participantRepository.save(participant), ParticipantDto.class);
-    }
-
 
     @PreAuthorize("hasRole('ROLE_USER')")
     @DeleteMapping
